feat: add matcher to unlock as well
